models and test_api updated to use Decimal

+++ app/models.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from typing import Optional
+from decimal import Decimal
 from pydantic import BaseModel
 
 
@@ -9,9 +10,9 @@ class Trade(BaseModel):
     account_id: int
     symbol: str
     side: str  # "BUY" or "SELL"
-    quantity: float
-    price: float
-    notional: float
+    quantity: Decimal
+    price: Decimal
+    notional: Decimal
     timestamp: Optional[datetime] = None
 
 
@@ -19,33 +20,11 @@ class Liquidation(BaseModel):
     """Model for liquidation events stored in Postgres"""
     id: Optional[int] = None
     account_id: int
-    equity: float
-    maintenance_margin: float
+    equity: Decimal
+    maintenance_margin: Decimal
     reason: str
     timestamp: Optional[datetime] = None
 
 
 # SQL table creation queries
-# TRADES_TABLE_SQL = """
-# CREATE TABLE IF NOT EXISTS trades (
-#     id SERIAL PRIMARY KEY,
-#     account_id INTEGER NOT NULL,
-#     symbol VARCHAR(20) NOT NULL,
-#     side VARCHAR(4) NOT NULL,
-#     quantity DECIMAL(20, 8) NOT NULL,
-#     price DECIMAL(20, 8) NOT NULL,
-#     notional DECIMAL(20, 8) NOT NULL,
-#     timestamp TIMESTAMP NOT NULL DEFAULT NOW()
-# );
-# """
 
-# LIQUIDATIONS_TABLE_SQL = """
-# CREATE TABLE IF NOT EXISTS liquidations (
-#     id SERIAL PRIMARY KEY,
-#     account_id INTEGER NOT NULL,
-#     equity DECIMAL(20, 8) NOT NULL,
-#     maintenance_margin DECIMAL(20, 8) NOT NULL,
-#     reason TEXT,
-#     timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-# );
-# """
